requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
flask-migrate==4.0.7
pymysql==1.1.1